requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
comet-ml==3.35.3
dvc==3.20.0
dvc-s3==3.0.1
PyYAML==6.0.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
flake8==6.1.0
pylint==2.17.5
black==23.7.0
requests==2.31.0
python-dateutil==2.8.2
colorlog==6.7.0
Werkzeug==2.3.7
